refactor: remove one layer (metrics), simplifying the call to qa_eval_engine
